+++ mayan/apps/documents/workflow_actions.py
@@ -1,0 +1,12 @@
+from __future__ import absolute_import, unicode_literals
+
+from django.utils.translation import ugettext_lazy as _
+
+from mayan.apps.document_states.classes import WorkflowAction
+
+
+class TrashDocumentAction(WorkflowAction):
+    label = _('Send document to trash')
+
+    def execute(self, context):
+        context['document'].delete()